Remove the deprecated function calls on OS X 10.5.
Because CoreAudio now uses separate handles to ioprocs, instead of the
function pointer, we perform a nasty hack to make this code still work
on OS X 10.4, by just using the new aprocid everywhere, but defining it
to the function name on OS X 10.4. I'm not sure whether we should check
for the OS X version like that.